requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
djangorestframework==3.14.0
flake8==6.0.0
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
